+++ src/histogram_class.cpp
@@ -8,7 +8,7 @@
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
@@ -20,21 +20,20 @@
    histogram_class.cpp contains the function declarations for the Gaussian Histogram Fitter
    See  histogram.h for the class definition
    1) histogram(double* data, int size);
-        Pass the 1d data array and the number of elements it contains.
-        The class will ask for a user input for histogram bin width and output the histgram data and fit to Hist.txt
+		Pass the 1d data array and the number of elements it contains.
+		The class will ask for a user input for histogram bin width and output the histgram data and fit to Hist.txt
    2) histogram(double* data, int size, double binsize);
-        Pass the 1d data array, the number of elements the array contains, and the histogram bin width.
-        The class will output  the histgram data and fit to Hist.txt
+		Pass the 1d data array, the number of elements the array contains, and the histogram bin width.
+		The class will output  the histgram data and fit to Hist.txt
    3) histogram(double* data, int size, char* fname);
-        Pass the 1d data array, the number of elements the array contains, and a file name to output the data and fit.
-        The class will ask for a user input for histogram bin width and output the fit and data to the passed file name.
+		Pass the 1d data array, the number of elements the array contains, and a file name to output the data and fit.
+		The class will ask for a user input for histogram bin width and output the fit and data to the passed file name.
    4) histogram(double* data, int size, double binsize, char* fname);
-        Pass the 1d data array, the number of elements the array contains, the histogram bin width, and a file name to output the data and fit.
-        The class will utput the fit and data to the passed file name.
+		Pass the 1d data array, the number of elements the array contains, the histogram bin width, and a file name to output the data and fit.
+		The class will utput the fit and data to the passed file name.
 
-        Jonathan Petticrew, University of Sheffield, 2017.
+		Jonathan Petticrew, University of Sheffield, 2017.
  */
-
 #include "histogram.h"
 #include <stdio.h>
 #include <math.h>
@@ -130,7 +129,7 @@ void histogram::max_min(){
 
 //Bins the DataSet and calculates mean and standard deviation
 void histogram::binner(){
-	bins=(max-min)/binsize+3;
+	bins=(int)((max-min)/binsize+3);
 	binEdges = new double[bins+1];
 	binCenters = new double[bins];
 	binValues = new double[bins];
